Main components added
