+++ frontend/src/components/Home/Header/WelcomeSection.css
@@ -1,0 +1,19 @@
+.welcome-section {
+  position: absolute;
+  left: 230px;
+  top: 12%;
+  display: flex;
+  align-items: center;
+}
+
+.welcome-text {
+  font-weight: 300;
+  font-size: 25px;
+  color: #1e1e1e;
+}
+
+.user-name {
+  font-weight: bold;
+  font-size: 25px;
+  color: #1e1e1e;
+}
