+++ frontend/src/components/Home/LeftSection/BottomSection.css
@@ -1,0 +1,8 @@
+.bottom-section {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 30px;
+  margin-top: 30px;
+  width: 95%;
+  height: calc(500px - 280px - 24px);
+}
