+++ frontend/src/components/Home/LeftSection/GetStartedCard.css
@@ -1,0 +1,31 @@
+.get-started-card {
+  background: white;
+  border-radius: 30px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
+  width: 85%;
+  padding: 20px 30px;
+  height: 50%;
+}
+
+.get-started-card h2 {
+  font-size: 20px;
+  font-weight: 600;
+  margin-top: 0;
+  color: #2f3c50;
+}
+
+.get-started-card p {
+  color: #3c4049;
+  font-weight: 400;
+  text-align: left;
+  font-size: 14px;
+  margin-bottom: 16px;
+}
+
+.button-stack {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 8px;
+  height: 55%;
+}
