--- frontend/src/pages/home/Home.css
@@ -1,335 +1,0 @@
-/* Shared Card Styles */
-.logbooks-card,
-.get-started-card,
-.recent-activity-card {
-  background: white;
-  border-radius: 30px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
-}
-
-.logbooks-card,
-.get-started-card {
-  width: 85%;
-}
-
-/* Shared Heading Styles */
-.get-started-card h2,
-.card-header h2,
-.activity-header h3,
-.shop-books-card h3 {
-  font-size: 20px;
-  font-weight: 600;
-  margin-top: 0;
-}
-
-.get-started-card h2,
-.card-header h2,
-.activity-header h3 {
-  color: #2F3C50;
-}
-
-.shop-books-card h3 {
-  color: white;
-}
-
-/* Lift-on-Hover Effect */
-.shop-books-card,
-.recent-activity-card {
-  transition: transform 0.2s ease;
-}
-
-.shop-books-card:hover,
-.recent-activity-card:hover {
-  transform: translateY(-2px);
-}
-
-/* Chevron Icon Styles */
-.chevron-icon {
-  width: 20px;
-  height: 20px;
-}
-
-.shop-books-card .chevron-icon {
-  color: white;
-  margin-top: 4px;
-}
-
-.activity-header .chevron-icon {
-  color: black;
-  margin-top: -20px;
-}
-
-/* Dashboard Container */
-.dashboard-container {
-  padding: 16px;
-  margin-left: 65px;
-  max-width: 1400px;
-}
-
-/* Welcome Section */
-.welcome-section {
-  position: absolute;
-  left: 230px;
-  top: 12%;
-  display: flex;
-  align-items: center;
-}
-
-.welcome-text {
-  font-weight: 300;
-  font-size: 25px;
-  color: #1E1E1E;
-}
-
-.user-name {
-  font-weight: bold;
-  font-size: 25px;
-  color: #1E1E1E;
-}
-
-/* Content Grid */
-.content-grid {
-  display: grid;
-  grid-template-columns: 1fr 1.2fr;
-  gap: 24px;
-  margin-top: 5%;
-  height: auto;
-}
-
-/* Logbooks Card */
-.logbooks-card {
-  padding: 24px;
-  height: 93%;
-}
-
-/* Get Started Card */
-.get-started-card {
-  padding: 20px 30px;
-  height: 50%;
-}
-
-.get-started-card p {
-  color: #3c4049;
-  font-weight: 400;
-  text-align: left;
-  font-size: 14px;
-  margin-bottom: 16px;
-}
-
-/* Card Header */
-.card-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-left: 30px;
-}
-
-.card-header h2 {
-  margin-bottom: 8px;
-}
-
-/* Button Stack */
-.button-stack {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 8px;
-  height: 55%;
-}
-
-/* View More Button */
-.view-more-btn {
-  padding: 6px 12px;
-  border-radius: 20px;
-  border: 1px solid #9AB0E1;
-  background: #F7FAFF;
-  color: #4F607E;
-  font-size: 15px;
-  font-weight: bold;
-  cursor: pointer;
-  margin-right: 30px;
-}
-
-/* Progress List */
-.progress-list {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin: 0 auto;
-  gap: 20px;
-  width: 80%;
-}
-
-/* Progress Item */
-.progress-item {
-  width: 100%;
-  padding: 12px;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: background-color 0.2s ease;
-  background-color: transparent;
-}
-
-.progress-item:hover {
-  background-color: #E7EDF8;
-}
-
-.progress-item:hover .progress-bar {
-  background: #FBFCFE;
-}
-
-.progress-item:hover .progress-info {
-  color: #244B94;
-}
-
-/* Progress Info */
-.progress-info {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 6px;
-  font-size: 15px;
-}
-
-/* Progress Bar */
-.progress-bar {
-  height: 8px;
-  background: #E7EDF8;
-  border-radius: 4px;
-  overflow: hidden;
-}
-
-.progress-fill {
-  height: 100%;
-  background: #244B94;
-  border-radius: 4px;
-}
-
-/* Logbooks Image */
-.logbooks-image {
-  width: 100%;
-  max-width: 280px;
-  margin: 30px auto;
-  display: block;
-}
-
-/* Bottom Section */
-.bottom-section {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 30px;
-  margin-top: 30px;
-  width: 95%;
-  height: calc(500px - 280px - 24px);
-}
-
-/* Shop Books Card */
-.shop-books-card {
-  background: linear-gradient(232.43deg, #6C8BD3 9.4%, #244B94 90.17%);
-  border-radius: 30px;
-  padding: 20px;
-  color: white;
-  display: flex;
-  justify-content: space-between;
-  cursor: pointer;
-  height: 83%;
-  position: relative;
-  overflow: hidden;
-}
-
-.shop-books-card::after {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: radial-gradient(
-    circle at bottom,
-    rgba(255, 255, 255, 0) 30%,
-    rgba(255, 255, 255, 0.2) 100%
-  );
-  pointer-events: none;
-}
-
-/* Shop Books Image */
-.shop-books-image {
-  position: absolute;
-  top: 43%;
-  left: 1px;
-  right: 1px;
-  width: 250px;
-  height: auto;
-  object-fit: contain;
-  z-index: 1;
-}
-
-/* Recent Activity Card */
-.recent-activity-card {
-  padding: 20px;
-  height: 83%;
-  max-height: 300px;
-  overflow-y: auto;
-}
-
-/* Activity Header */
-.activity-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-/* Activity List */
-.activity-list {
-  display: flex;
-  flex-direction: column;
-  margin-top: -10px;
-}
-
-/* Activity Item */
-.activity-item {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  border-bottom: 1px solid #e0e0e0;
-}
-
-/* Activity Info */
-.activity-info {
-  display: flex;
-  gap: 5px;
-  color: #666;
-  font-size: 12px;
-}
-
-.activity-info h2 {
-  font-weight: 500;
-  font-size: 12px;
-  color: #2F3C50;
-  opacity: 0.8;
-}
-
-.activity-info h3 {
-  font-weight: 700;
-  font-size: 12px;
-  color: #2E4A8F;
-  margin-top: 10px;
-}
-
-/* Activity Time */
-.activity-time {
-  margin-left: 20px;
-}
-
-.activity-time h3 {
-  font-weight: 400;
-  font-size: 12px;
-  color: #A1B4DE;
-}
-
-/* Time Icon */
-.time-icon {
-  color: #92a4cb;
-  width: 14px;
-  height: 14px;
-}
