+++ frontend/src/pages/home/Home.jsx
@@ -1,174 +1,10 @@
-import { useState } from "react";
-import { CLButtonSecondary } from "../../components/Buttons/CLButtons";
-import { useNavigate } from "react-router-dom";
-import LogBooks from "../../assets/images/logbooks.png";
-import ShopLogBooks from "../../assets/images/ShopLogBooks.png";
-import { ChevronRightIcon, ClockIcon } from "@heroicons/react/24/outline";
-import "./Home.css";
-import { useAuth } from "../../contexts/AuthContext";
 import { NavContentWrapper } from "../../components/NavContentWrapper/NavContentWrapper";
-import { NewLogModal } from "../../components/NewLogModal/NewLogModal";
+import MainContent from "../../components/Home/LeftSection/MainContent";
 
 export default function Home() {
   return (
     <NavContentWrapper>
       <MainContent />
     </NavContentWrapper>
-  );
-}
-
-function MainContent() {
-  const navigate = useNavigate();
-  const [setSelectedLog] = useState(null);
-  const { session } = useAuth();
-
-  const handleAddLogbook = () => {
-    navigate("/newLog");
-  };
-
-  const handleViewHistory = () => {
-    navigate("/history");
-  };
-
-  const recentActivities = [
-    {
-      id: 1,
-      action: "Added Log",
-      logName: "mylogexample",
-      time: "1d",
-    },
-    {
-      id: 2,
-      action: "Added Log",
-      logName: "mylogexample",
-      time: "1d",
-    },
-    {
-      id: 3,
-      action: "Added Log",
-      logName: "mylogexample",
-      time: "1d",
-    },
-    // Add more items as needed
-  ];
-
-  const progressItems = [
-    {
-      id: 1,
-      name: "Adult Cardiac 2025",
-      progress: 65,
-    },
-    {
-      id: 2,
-      name: "Adult Cardiac 2025",
-      progress: 65,
-    },
-    {
-      id: 3,
-      name: "Adult Cardiac 2025",
-      progress: 65,
-    },
-    // Add more items as needed
-  ];
-
-  return (
-    <div className="dashboard-container">
-      <div className="welcome-section">
-        <h1>
-          <span className="welcome-text">Welcome back, </span>
-          <span className="user-name">
-            {session?.user?.user_metadata?.first_name || "User"}
-          </span>
-        </h1>
-      </div>
-
-      <div className="content-grid">
-        <div>
-          <div className="get-started-card">
-            <h2>Get Started</h2>
-            <p>
-              Convert handwritten clinical logs to standardized excel templates
-              with just a click of a button!
-            </p>
-            <div className="button-stack">
-              <NewLogModal />
-              <CLButtonSecondary onClick={handleAddLogbook} width={"230px"}>
-                Add Logbook
-              </CLButtonSecondary>
-              <CLButtonSecondary onClick={handleViewHistory} width={"230px"}>
-                View Log History
-              </CLButtonSecondary>
-            </div>
-          </div>
-
-          <div className="bottom-section">
-            <a href="/shop-logbooks" className="shop-books-link">
-              <div className="shop-books-card">
-                <h3>Shop Log Books</h3>
-                <ChevronRightIcon className="chevron-icon" />
-                <img
-                  src={ShopLogBooks}
-                  alt="Shop Log Books"
-                  className="shop-books-image"
-                />
-              </div>
-            </a>
-
-            <a href="/recent-activity" className="recent-activity-link">
-              <div className="recent-activity-card">
-                <div className="activity-header">
-                  <h3>Recent Activity</h3>
-                  <ChevronRightIcon className="chevron-icon" />
-                </div>
-                <div className="activity-list">
-                  {recentActivities.map((activity) => (
-                    <div key={activity.id} className="activity-item">
-                      <div className="activity-info">
-                        <h2>{activity.action}:</h2>
-                        <h3 className="log-name">{activity.logName}</h3>
-                      </div>
-                      <div className="activity-time">
-                        <h3>{activity.time}</h3>
-                      </div>
-                      <ClockIcon className="time-icon" />
-                    </div>
-                  ))}
-                </div>
-              </div>
-            </a>
-          </div>
-        </div>
-
-        <div className="logbooks-card">
-          <div className="card-header">
-            <h2>Log Books</h2>
-            <button className="view-more-btn">View more</button>
-          </div>
-          <div className="logbooks-content">
-            <img src={LogBooks} alt="Logbooks" className="logbooks-image" />
-            <div className="progress-list">
-              {progressItems.map((item) => (
-                <div
-                  key={item.id}
-                  className="progress-item"
-                  onClick={() => setSelectedLog(item.id)}
-                >
-                  <div className="progress-info">
-                    <span>{item.name}</span>
-                    <span>{item.progress}%</span>
-                  </div>
-                  <div className="progress-bar">
-                    <div
-                      className="progress-fill"
-                      style={{ width: `${item.progress}%` }}
-                    ></div>
-                  </div>
-                </div>
-              ))}
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
   );
 }
